@@ -2,26 +2,30 @@
   "name": "portfolio",
   "version": "0.1.0",
   "private": true,
+  "homepage": "http://github.com/krish4uu/kashyap-prajapati",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.6",
-    "@testing-library/react": "^11.2.2",
-    "@testing-library/user-event": "^12.6.0",
-    "axios": "^0.21.1",
-    "bootstrap": "^4.5.3",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.4.0",
-    "react-dom": "^17.0.1",
-    "react-github-calendar": "^2.0.2",
-    "react-icons": "^4.3.1",
-    "react-parallax-tilt": "^1.4.68",
-    "react-particles-js": "^3.4.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "typewriter-effect": "^2.17.0",
-    "web-vitals": "^0.2.4"
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^1.4.0",
+    "bootstrap": "^5.3.1",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.8.0",
+    "react-dom": "^18.2.0",
+    "react-github-calendar": "^4.0.1",
+    "react-icons": "^4.10.1",
+    "react-parallax-tilt": "^1.7.154",
+    "react-router-dom": "^6.14.2",
+    "react-scripts": "5.0.1",
+    "react-tsparticles": "^2.11.0",
+    "tsparticles": "^2.11.1",
+    "typewriter-effect": "^2.20.1",
+    "web-vitals": "^3.4.0"
   },
   "scripts": {
     "start": "react-scripts start",
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -43,5 +47,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "gh-pages": "^5.0.0"
   }
 }
