@@ -1,7 +1,6 @@
 import React from "react";
-import Particle from "./Particle";
 function Pre(props) {
-  return <div id={props.load ? "preloader" : "preloader-none"}><Particle /></div>;
+  return <div id={props.load ? "preloader" : "preloader-none"}></div>;
 }
 
 export default Pre;
